requirements.txt
anthropic>=0.40.0
pdfplumber>=0.11.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
llm-snowglobe
tiktoken>=0.8.0
